requirements.txt
datasets
openai
numpy
tqdm
huggingface-hub
python-dotenv
